pyproject.toml
[project]
name = "rag-knowledge-base"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "beanie>=2.0.0",
    "celery>=5.5.3",
    "content-core[docling]>=1.0.2",
    "fastapi>=0.116.1",
    "google-api-core>=2.25.1",
    "google-cloud-storage>=3.3.0",
    "gunicorn>=23.0.0",
    "langchain>=0.3.27",
    "langchain-community>=0.3.27",
    "langchain-google>=0.1.1",
    "langchain-google-community>=2.0.7",
    "langchain-openai>=0.3.30",
    "loguru>=0.7.3",
    "motor>=3.7.1",
    "notebook>=7.4.5",
    "psycopg2>=2.9.10",
    "psycopg2-binary>=2.9.10",
    "pydantic>=2.11.7",
    "pymongo>=4.14.0",
    "pymysql>=1.1.1",
    "pypdf2>=3.0.1",
    "python-docx>=1.2.0",
    "python-dotenv>=1.1.1",
    "redis>=6.4.0",
    "unicorn>=2.1.3",
    "unstructured>=0.18.13",
]
